requirements.txt
snowflake-connector-python
python
dbt